+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RepositorySearchDelegateProtocol` implementation.
 - `RepositoryListViewController` implementation.
 - `RepositoryListTableViewSection` implementation.
+- `Builder` protocol implementation.
+- `ModuleBuilder` implementation.
 
 ## [0.1.0] - 2023-05-02